Issue 18

+++ killer.sh
@@ -8,7 +8,12 @@ Usage: $0 <cluster-name>
 
 Deletes every node of the form <cluster-name>-node*
 EOF
+    exit 1
 fi
 
 nodes=$(gcloud compute instances list | grep ^$1-node | awk '{ print $1 }')
+if [[ ! $nodes ]]; then
+    echo "Could not find cluster $1"
+    exit 1
+fi
 gcloud compute instances delete -q $nodes
